sync: Use options from the txt file
The requirements.txt file might have index or find-links options which
affects finding of the installable packages.  Parse those options from
the txt file before doing the sync.

Utilize the pip options parsing code from the compile command which is
now moved to a separate file.
